Some cleaning and moving classes around
