+++ src/main/lombok/eu/coinform/gateway/controller/forms/LogoutForm.java
@@ -1,4 +1,4 @@
-package eu.coinform.gateway.controller;
+package eu.coinform.gateway.controller.forms;
 
 import lombok.Data;
 
@@ -6,9 +6,9 @@ import javax.validation.constraints.Email;
 import javax.validation.constraints.NotEmpty;
 
 @Data
-public class RegisterForm {
+public class LogoutForm {
     @Email
-    String email;
+    private String email;
     @NotEmpty
-    String password;
+    private String password;
 }
