+++ src/main/lombok/eu/coinform/gateway/db/UserDbAuthenticationFilter.java
@@ -1,14 +1,12 @@
 package eu.coinform.gateway.db;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import eu.coinform.gateway.controller.LoginForm;
+import eu.coinform.gateway.controller.forms.LoginForm;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.security.authentication.AuthenticationManager;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.AuthenticationException;
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.web.authentication.AbstractAuthenticationProcessingFilter;
 import org.springframework.security.web.util.matcher.RegexRequestMatcher;
@@ -19,8 +17,6 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
 import java.io.IOException;
-import java.util.LinkedList;
-import java.util.List;
 
 
 @Slf4j
